feat: Scaffold initial Flutter project for App Store frontend
This commit introduces the initial scaffolding for the Flutter-based frontend of the internal app store. It replaces the boilerplate Flutter application with a structured, feature-rich starting point.

It includes the following:
- A modular project structure with directories for features, services, and models.
- All required dependencies, including Riverpod, Dio, flutter_dotenv, and others.
- Basic UI pages for Login, App List, and an Admin Dashboard with an upload form.
- Services for handling API communication and authentication, including a Dio client with a JWT interceptor.
- Environment configuration setup using a .env file for the backend URL, with the .env file added to .gitignore.
- A comprehensive README with setup and usage instructions.
