--- app_store_frontend/pubspec.yaml
@@ -1,29 +1,0 @@
-name: app_store_frontend
-description: An internal app store for Android and iOS apps.
-publish_to: 'none'
-version: 1.0.0+1
-
-environment:
-  sdk: '>=3.0.0 <4.0.0'
-
-dependencies:
-  flutter:
-    sdk: flutter
-  flutter_riverpod: ^2.3.6
-  dio: ^5.1.1
-  flutter_dotenv: ^5.0.2
-  shared_preferences: ^2.0.15
-  flutter_secure_storage: ^8.0.0
-  file_picker: ^5.2.5
-  url_launcher: ^6.1.10
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^2.0.0
-
-flutter:
-  uses-material-design: true
-
-  assets:
-    - .env
