+++ lib/src/models/app_model.dart
