+++ lib/src/services/auth_service.dart
