+++ pubspec.yaml
@@ -34,6 +34,13 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
+  flutter_riverpod: ^2.3.6
+  dio: ^5.1.1
+  flutter_dotenv: ^5.0.2
+  shared_preferences: ^2.0.15
+  flutter_secure_storage: ^8.0.0
+  file_picker: ^5.2.5
+  url_launcher: ^6.1.10
 
 dev_dependencies:
   flutter_test:
@@ -58,9 +65,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - .env
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
